mirage: check in any generated .xl.in file
mirage 2.7.0 will generate an .xl.in file as part of the configure
phase. This file contains the virtual hardware configuration needed
by the kernel, with holes where the network bridges, disk images
and memory settings should go.

This patch ensures that any generated .xl.in file is checked into
the binary repo alongside the kernel that it describes.

Deployment scripts should use the .xl.in files as a template, so there
is no unexpected breakage if (for example) the virtual disk device
changes from xvda to xvdb, or if the VM is booted HVM rather than PV.

This is part of [mirage/mirage#443]

Signed-off-by: David Scott <dave.scott@docker.com>