Add support for BASE_REMOTE_BRANCH
So far, .travis-docker.sh uses branch "master" of any BASE_REMOTE that
is being supplied. This commit adds support to optionally use a
different branch which is supplied via BASE_REMOTE_BRANCH.

The motivation is to enable versioning of the Opam repository that is
being used. When such a repository is used to describe the universe of
packages for a product, it is natural to evolve it and the desire to
refer to a past status arises naturally.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>