Run tests on revdeps
There does not seem to be a way to run reverse-dependencies tests
currently. We would like to do that for ppx_deriving CI (see
https://github.com/ocaml-ppx/ppx_deriving/pull/227).

In this commit, I propose to run tests on reverse-dependencies when
`TESTS=true` (the default) and `REVDEPS` is `true` or lists some
reverse dependencies to test (no by default).